requirement.txt
altgraph==0.17.3
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.0.1
dnspython==2.3.0
future==0.18.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
joblib==1.2.0
numpy==1.24.1
packaging==23.0
pandas==1.5.3
pefile==2022.5.30
Pillow==9.4.0
pyaes==1.6.1
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.15
pymongo==4.3.3
PyPDF2==3.0.1
Pyrogram==2.0.99
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram==0.16.0
python-telegram-bot==20.0
pytz==2022.7.1
pywin32-ctypes==0.2.0
requests==2.28.2
rfc3986==1.5.0
scikit-learn==1.2.1
scipy==1.10.0
six==1.16.0
sniffio==1.3.0
TgCrypto==1.2.5
threadpoolctl==3.1.0
urllib3==1.26.14
